+++ test/templating/react.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { withTranslation, Trans, Interpolate } from 'react-i18next'
 
+// These will have namespace "react" even though it comes before withtranslation during parsing
 const bar = () => (
   <div>
     <span><Trans i18nKey="bar"></Trans></span>
@@ -37,6 +38,7 @@ class Test extends React.Component {
         <Trans>
           don't split {{ on: this }}
         </Trans>
+        <Trans i18nKey="override-default" defaults="default override">ignore me</Trans>
       </div>
     )
   }